SPLIT not WORKING RETARD
