+++ app/src/main/AndroidManifest.xml
@@ -57,7 +57,7 @@
             </intent-filter>
 
         </activity>
-        <activity android:name=".SettingsActivity" />
+        <activity android:name=".settings.SettingsActivity" />
 
 
     </application>
