+++ app/src/main/java/com/djangofiles/djangofiles/settings/SettingsActivity.kt
@@ -1,0 +1,15 @@
+package com.djangofiles.djangofiles.settings
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class SettingsActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        supportFragmentManager
+            .beginTransaction()
+            .replace(android.R.id.content, SettingsFragment())
+            .commit()
+    }
+}
